+++ src/features/des/cat-adi3/UnsucessfulTests/UnsuccessfulTestRiskManagementCatADI3.feature
@@ -1,3 +1,4 @@
+@test
 Feature: Cat-ADI3 Successful feature
 
   Scenario: Successful Grade B ADI3 test
